pyproject.toml
# A build system is required to convert your code into a distributable package.
# setuptools is the oldest and most common build tool, but we also like Poetry
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "msfocr"
version = "1.1.0"
description = "The backend to perform OCR on Doctors without Borders tally sheets."
readme = "README.md"

# What version of python does your library work with?
requires-python = "<3.13,>3.7"

# Metadata about your package in case you upload it to PYPI
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

# All the dependencies needed for running your module go here
dependencies = [
    "numpy",
    "pandas",
    "python-Levenshtein",
    "requests"
]

[project.optional-dependencies]
# Extra dependencies only needed for running tests go here
test = [
    "pytest",
    "pytest-datadir",
    "requests_mock",
    ]

# Dependencies only needed to run the streamlit app go here
app = [
    "openai",
    "streamlit",
    "simpleeval"
    ]

app-doctr = [
    "img2table",
    "python-doctr",
    "simpleeval",
    "streamlit",
    "torch",
    "torchvision",
]

# Dependencies that are useful only to developers, like an autoformatter and support for visualizations in jupyter notebooks go here
dev = [
    "azure-common==1.1.28",
    "azure-core==1.26.1",
    "azure-identity==1.12.0",
    "azure-keyvault==4.2.0",
    "azure-keyvault-secrets==4.6.0",
    "azure-storage-blob==12.14.1",
    "ruff",
    "jupyter",
    "matplotlib",
    "notebook",
    "seaborn",
    "sphinx",
]

# If your project contains scripts you'd like to be available command line, you can define them here.
# The value must be of the form "<package_name>:<module_name>.<function>"
[project.scripts]
